Fix giant broadcasting bug

+++ src/game/server/teams.cpp
@@ -257,7 +257,7 @@ bool CGameTeams::TeamFinished(int Team)
 
 int64_t CGameTeams::TeamMask(int Team, int ExceptID, int Asker)
 {
-	int Mask = 0;
+	int64_t Mask = 0;
 
 	for (int i = 0; i < MAX_CLIENTS; ++i)
 	{
